refactor(search-view): extract useSearchExecution and useUrlSync hooks
